Remove TPC-H connector dependency on DuckDB
