+++ velox/duckdb/README.md
@@ -15,10 +15,6 @@ Then copy the generated files to velox/external/duckdb:
     export VELOX_PATH="<path/to/velox>"
     rsync -vrh src/amalgamation/duckdb* ${VELOX_PATH}/velox/external/duckdb/
 
-We also maintain a copy of TPC-H dataset generators that need to be updated:
-
-    rsync -vrh --exclude={'CMakeLists.txt','LICENSE','*.py','dbgen/queries','dbgen/answers'} extension/tpch/ ${VELOX_PATH}/velox/external/duckdb/tpch/
-
 After the new files are copied, ensure that the new code compiles and that it
 doesn't break any tests. Velox relies on many internal APIs, so there is a good
 chance that this will not work out-of-the-box and that you will have to dig in
