+++ velox/duckdb/conversion/DuckWrapper.cpp
@@ -17,7 +17,6 @@
 #include "velox/common/base/BitUtil.h"
 #include "velox/duckdb/conversion/DuckConversion.h"
 #include "velox/external/duckdb/duckdb.hpp"
-#include "velox/external/duckdb/tpch/include/tpch-extension.hpp"
 #include "velox/vector/FlatVector.h"
 
 namespace facebook::velox::duckdb {
@@ -63,7 +62,6 @@ DuckDBWrapper::DuckDBWrapper(core::ExecCtx* context, const char* path)
     : context_(context) {
   db_ = std::make_unique<DuckDB>(path);
   connection_ = std::make_unique<Connection>(*db_);
-  db_->LoadExtension<::duckdb::TPCHExtension>();
 }
 
 DuckDBWrapper::~DuckDBWrapper() {}
