+++ velox/external/duckdb/CMakeLists.txt
@@ -9,8 +9,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-add_subdirectory(tpch)
-
 add_compile_definitions(DISABLE_DUCKDB_REMOTE_INSTALL)
 
 add_library(
