+++ velox/tpch/gen/CMakeLists.txt
@@ -11,9 +11,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+add_subdirectory(dbgen)
+
 add_library(velox_tpch_gen TpchGen.cpp DBGenIterator.cpp)
 
-target_link_libraries(velox_tpch_gen velox_memory velox_vector dbgen duckdb)
+target_include_directories(velox_tpch_gen PRIVATE dbgen/include)
+
+target_link_libraries(velox_tpch_gen velox_memory velox_vector velox_dbgen)
 
 if(${VELOX_BUILD_TESTING})
   add_subdirectory(tests)
