+++ velox/tpch/gen/DBGenIterator.h
@@ -19,8 +19,8 @@
 #include <memory>
 #include <mutex>
 
-#include "velox/external/duckdb/tpch/dbgen/include/dbgen/dss.h"
-#include "velox/external/duckdb/tpch/dbgen/include/dbgen/dsstypes.h"
+#include <velox/tpch/gen/dbgen/include/dbgen/dss.h>
+#include <velox/tpch/gen/dbgen/include/dbgen/dsstypes.h>
 
 namespace facebook::velox::tpch {
